pyproject.toml
[project]
name = "consumer-trend-agent"
description = "Automatic Consumer Trend Analysis Multi-Agent System"
version = "4.0.0"
authors = [
    { name = "Trend Analysis Team", email = "dev@example.com" }
]
license = { file = "LICENSE" }
readme = "README.md"
requires-python = ">= 3.11, < 3.13"

dependencies = [
    # Web Framework
    "fastapi==0.115.7",
    "uvicorn[standard]==0.35.0",
    "starlette>=0.27.0",

    # Data Validation & Configuration
    "pydantic==2.11.7",
    "pydantic-settings>=2.0.0",
    "PyYAML>=6.0",

    # HTTP Requests
    "requests==2.32.4",
    "aiohttp==3.12.15",
    "httpx==0.28.1",
    "urllib3>=2.0.0",

    # LLM & AI Providers
    "openai",
    "anthropic",
    "google-generativeai==0.8.5",

    # LangChain & LangGraph
    "langchain==0.3.26",
    "langchain-community==0.3.27",
    "langchain-core>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.3.0",
    "langchain-google-genai>=2.0.0",

    # MCP (Model Context Protocol)
    "mcp>=1.0.0",
    "langchain-mcp-adapters>=0.1.0",

    # Vector Database
    "pinecone>=5.0.0",

    # Embedding Providers
    "voyageai>=0.2.0",

    # NLP & Text Processing
    "textblob>=0.17.0",
    "nltk>=3.8.0",

    # Data Processing
    "pandas==2.2.3",
    "numpy",
    "scikit-learn>=1.3.0",

    # RSS Feed Parsing
    "feedparser>=6.0.0",

    # Utilities
    "python-dotenv",
    "tiktoken",
    "psutil>=5.9.0",

    # WebSocket support
    "websockets>=12.0",

    # Redis (optional)
    "redis>=5.0.0",

    # Monitoring (optional)
    "prometheus-client>=0.19.0",

    # Type stubs for better IDE support
    "types-requests>=2.31.0",
    "types-PyYAML>=6.0.0",
    "types-redis>=4.6.0",
    "types-psutil>=5.9.0",

    # Testing
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
]

[project.optional-dependencies]
dev = [
    "black==25.1.0",
    "pytest-cov>=4.0.0",
    "ipython",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.black]
line-length = 100
target-version = ['py311']
